Refactored eldritch_run into eldritch lib.
