+++ implants/golem/src/main.rs
@@ -5,54 +5,7 @@ use clap::{Command, Arg};
 use tokio::task;
 use std::fs;
 
-use starlark::{starlark_module};
-use starlark::environment::{GlobalsBuilder, Globals, Module};
-use starlark::syntax::{AstModule, Dialect};
-use starlark::eval::Evaluator;
-use starlark::values::Value;
-
-use eldritch::file::FileLibrary;
-use eldritch::process::ProcessLibrary;
-use eldritch::sys::SysLibrary;
-
-fn eldritch_run(tome_filename: String, tome_contents: String) -> Result<String, golem::Error> {
-    let ast: AstModule = AstModule::parse(
-        &tome_filename, 
-        tome_contents.as_str().to_owned(), 
-        &Dialect::Standard
-    ).unwrap();
-
-    #[starlark_module]
-    fn eldritch(builder: &mut GlobalsBuilder) {
-        const file: FileLibrary = FileLibrary();
-        const process: ProcessLibrary = ProcessLibrary();
-        const sys: SysLibrary = SysLibrary();
-    }
-
-    let globals = GlobalsBuilder::extended().with(eldritch).build();
-    let module: Module = Module::new();
-
-    let mut eval: Evaluator = Evaluator::new(&module);
-
-    let res: Value = eval.eval_module(ast, &globals).unwrap();
-
-    println!("{:?}", res.unpack_str());
-
-    // let mut eld = Evaluator::eval_function(
-    //     &mut self, "dir(file)", positional, named
-    // );
-
-//     let mut a = Assert::new();
-//     a.globals_add(globals);
-//     a.all_true(
-//         r#"
-// dir(file) == ["append", "copy", "download", "exists", "hash", "is_dir", "mkdir", "read", "remove", "rename", "replace", "replace_all", "timestomp", "write"]
-// dir(process) == ["kill", "list", "name"]
-// dir(sys) == ["exec", "is_linux", "is_macos", "is_windows", "shell"]
-// "#,
-//     );
-    Ok("Ran tome".to_string())
-}
+use eldritch::eldritch_run;
 
 async fn interactive() -> Result<(), golem::Error> {
     Ok(())
@@ -63,7 +16,7 @@ async fn run(tome_path: String) -> Result<String, golem::Error> {
     // Read a tome script
     let tome_contents = fs::read_to_string(tome_path.clone())?;
     // Execute a tome script
-    let tome_results: String = eldritch_run(tome_path, tome_contents)?;
+    let tome_results: String = eldritch_run(tome_path, tome_contents).unwrap();
     // Return script output
     return Ok(tome_results)
 }
